Split up state into smaller modules
