+++ src/parsita/__init__.py
@@ -1,3 +1,4 @@
+from .exceptions import ParseError, RecursionError
 from .metaclasses import GeneralParsers, TextParsers, fwd
 from .parsers import (
     Parser,
@@ -18,4 +19,5 @@ from .parsers import (
     success,
     until,
 )
-from .state import Failure, ParseError, Reader, RecursionError, Result, SequenceReader, StringReader, Success
+from .reader import Reader, SequenceReader, StringReader
+from .result import Failure, Result, Success
