+++ src/parsita/parsers/_debug.py
@@ -2,7 +2,8 @@ __all__ = ["DebugParser", "debug"]
 
 from typing import Callable, Generic, Optional
 
-from ..state import Input, Output, Reader, State
+from ..reader import Reader
+from ..state import Input, Output, State
 from ._base import Parser
 from ._literal import lit
