+++ src/parsita/parsers/_optional.py
@@ -2,7 +2,8 @@ __all__ = ["OptionalParser", "opt"]
 
 from typing import Generic, List, Sequence, Union
 
-from ..state import Continue, Input, Output, Reader, State
+from ..reader import Reader
+from ..state import Continue, Input, Output, State
 from ._base import Parser
 from ._literal import lit
