+++ src/parsita/parsers/_repeated.py
@@ -2,7 +2,9 @@ __all__ = ["RepeatedOnceParser", "rep1", "RepeatedParser", "rep"]
 
 from typing import Generic, Optional, Sequence, Union
 
-from ..state import Continue, Input, Output, Reader, RecursionError, State
+from ..exceptions import RecursionError
+from ..reader import Reader
+from ..state import Continue, Input, Output, State
 from ._base import Parser
 from ._literal import lit
