+++ src/parsita/parsers/_success.py
@@ -2,7 +2,8 @@ __all__ = ["SuccessParser", "success", "FailureParser", "failure"]
 
 from typing import Any, Generic, NoReturn
 
-from ..state import Continue, Input, Output, Reader, State
+from ..reader import Reader
+from ..state import Continue, Input, Output, State
 from ._base import Parser
