+++ src/parsita/parsers/_until.py
@@ -2,7 +2,8 @@ __all__ = ["UntilParser", "until"]
 
 from typing import Any, Generic
 
-from ..state import Continue, Input, Output, Reader, State
+from ..reader import Reader
+from ..state import Continue, Input, Output, State
 from ._base import Parser
 from ._literal import lit
